Add markOrderNoteAsReadIfNeeded helper method
